requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
django-cors-headers
djangorestframework-simplejwt
celery[redis]
redis
pandas
python-dotenv
gunicorn
scikit-learn
numpy
django-filter
